+++ R/MDP.R
@@ -107,8 +107,6 @@ MDP <- function(states,
 
 
 #' @export
-
-#' @export
 print.MDP <- function(x, ...) {
   writeLines(paste(paste(class(x), collapse = ", "),
     "-",
@@ -125,7 +123,7 @@ print.MDP <- function(x, ...) {
   writeLines(sprintf("  Size: %d states / %d actions\n",
     length(x$states), length(x$actions)))
   
-  if (.solved_MDP(x))
+  if (is_solved_MDP(x))
     writeLines(c(
       "  Solved:",
       sprintf("    Solution converged: %s",
@@ -159,7 +157,10 @@ MDP2POMDP <- function(x) {
   x
 }
 
-.solved_MDP <- function(x, stop = FALSE) {
+#' @rdname MDP
+#' @param stop logical; stop with an error.
+#' @export
+is_solved_MDP <- function(x, stop = FALSE) {
   if (!inherits(x, "MDP"))
     stop("x needs to be a MDP object!")
   solved <- !is.null(x$solution)