+++ R/plot_belief_space.R
@@ -38,7 +38,7 @@
 #' plot_belief_space(sol, n = 10, sample = "random")
 #'
 #' # plot the belief points used by the grid-based solver
-#' plot_belief_space(sol, sample = sol$solution$belief_states)
+#' plot_belief_space(sol, sample = sol$solution$belief_points_solver)
 #'
 #' # plot different measures
 #' plot_belief_space(sol, what = "pg_node")
@@ -56,7 +56,7 @@
 #' plot_belief_space(sol, what = "reward", sample = "random", n = 1000)
 #'
 #' # plot the belief points used by the grid-based solver
-#' plot_belief_space(sol, sample = sol$solution$belief_states)
+#' plot_belief_space(sol, sample = sol$solution$belief_points_solver)
 #'
 #' # plot the belief points obtained using simulated trajectories with an epsilon-greedy policy.
 #' # Note that we only use n = 50 to save time.