+++ R/read_write_POMDP.R
@@ -23,6 +23,9 @@ format_fixed <- function(x, digits = 7, debug = "unknown") {
 #' is parsed using an experimental POMDP file parser. The parsed information can be used with auxiliary functions
 #' in this package that use fields like the transition matrix, the observation matrix and the reward structure.
 #'
+#' **Note:** The parser for POMDP files is experimental. Please report
+#' problems here: \url{https://github.com/mhahsler/pomdp/issues}.
+#'
 #' @family POMDP
 #'
 #' @param x an object of class [POMDP].