+++ R/reward.R
@@ -60,7 +60,7 @@ reward <- function(x, belief = NULL, epoch = 1) {
 #' @rdname reward
 #' @export
 reward_node_action <- function(x, belief = NULL, epoch = 1) {
-  .solved_POMDP(x)
+  is_solved_POMDP(x)
   
   if (is.null(belief))
     belief <- x$start